refactor(nats): split llm_adapter.py — extract generation into mixin
src/llmcli/nats/llm_adapter.py was 380 lines, over the 300-line
file_length quality gate. Extract HTTP generation logic into
src/llmcli/nats/_generation.py as GenerationMixin; the adapter shell
(NATS lifecycle, _ensure_model, heartbeat) stays in llm_adapter.py.

LlmNatsAdapter now inherits GenerationMixin before NatsAdapterBase.
The mixin accesses self._client, self._loaded_model and self._nc —
all instance state defined on LlmNatsAdapter — so no argument
threading is needed and method signatures stay identical.

Remove tools/file_exemptions.txt entry — both files now under the
gate (llm_adapter.py: 224 lines, _generation.py: 207 lines).
Tests pass without modification (283 passed).

Closes #26
